Updated Molecule Editor classes

+++ VParadigm/NTU_2014FYP_Common/gen/sg/edu/ntu/aalhossary/fyp2014/moleculeeditor/core/EvaluateUserAction.java
@@ -26,6 +26,7 @@ public class EvaluateUserAction {
 	public static final int DRAGATOMMINIMIZE = 6;
 	public static final int DRAGSELECTED = 7;
 	public static final int DRAGSELECTEDMINIMIZE = 8;
+	public static final int MINIMIZE = 9;
 	
 	public static void evaluateAction(UpdateRegistry updateReg, String script) {
 		System.out.println("Evaluating script: " + script);
@@ -50,11 +51,15 @@ public class EvaluateUserAction {
 			//System.out.println(viewer.getAtomPoint3f(10).x+  " "+ viewer.getAtomPoint3f(10).y + " " +viewer.getAtomPoint3f(10).z);
 			findSelectedAtom(updateReg, PASTE);
 			return;
+		case "minimizeonly":
+			System.out.println("Entering drag Atom function");
+			dragModeSetting(updateReg, MINIMIZE);
+			return;
 		case "dragatom":
 			System.out.println("Entering drag Atom function");
 			dragModeSetting(updateReg, DRAGATOM);
 			return;
-		case "dragatomminimize":
+		case "dragminimize":
 			dragModeSetting(updateReg, DRAGATOMMINIMIZE);
 			return;
 		case "dragselected":
@@ -70,32 +75,56 @@ public class EvaluateUserAction {
 
 	private static void dragModeSetting(UpdateRegistry updateReg, int dragMode) {
 			updateReg.dragMode = DRAGATOM;
-			if(dragMode==DRAGATOM || dragMode==DRAGSELECTED){
+			if(dragMode==MINIMIZE){
 				System.out.println("Evaluating set picking dragatom");
-				updateReg.viewer.evalString("set picking DRAGATOM");
+				updateReg.minimizeModel();
+			}
+			else if(dragMode==DRAGATOM){
+				updateReg.atomselected=true;
+				System.out.println("Evaluating set picking dragatom");
+				updateReg.getViewer().evalString("set picking DRAGATOM");
+			}
+			else if(dragMode==DRAGSELECTED){
+				updateReg.atomselected=false;
+				System.out.println("Evaluating set picking dragatom");
+				updateReg.getViewer().evalString("set picking DRAGATOM");
 			}
 			else if(dragMode==DRAGATOMMINIMIZE || dragMode==DRAGSELECTEDMINIMIZE){
+				updateReg.atomselected=true;
 				System.out.println("Evaluating set picking dragatom");
-				int model = updateReg.viewer.getDisplayModelIndex();
-				updateReg.viewer.evalString("set picking DRAGATOM");
-				Minimizer minimizer = new Minimizer();
-				minimizer.setupMinimizing(updateReg.modelList.get(model));
+				updateReg.getViewer().evalString("set picking DRAGATOM");
+				updateReg.setMinimizeMode(true);
 			}
 	}
 
 	public static void findSelectedAtom(UpdateRegistry updateReg, int mode) {
 		ArrayList<Atom> atomList = updateReg.getSelectedAtoms();
-		if(mode==DELETE || mode==CUT){
+		int[] atomNo = new int[atomList.size()];
+		for(int i=0;i<atomNo.length;i++){
+			atomNo[i] = atomList.get(atomNo.length-1-i).getAtomSeqNum();
+		}
+		int currentModelIndex = updateReg.getViewer().getDisplayModelIndex(); 
+		if(mode==DELETE ||mode==CUT||mode==COPY){
 			for(int i=0;i<atomList.size();i++){
 				String key = atomList.get(i).getModelName()+atomList.get(i).getAtomSeqNum();
-				int currentModelIndex = updateReg.viewer.getDisplayModelIndex(); 
-				updateReg.deleteAtoms(currentModelIndex, key, atomList.get(i).getAtomSeqNum());
+				if(mode==CUT || mode==COPY)
+					updateReg.copyAtoms(currentModelIndex, key);
+				if(mode==DELETE || mode==CUT)
+					updateReg.deleteAtoms(currentModelIndex, key, atomList.get(i).getAtomSeqNum());
 			}
 		}
 		else if(mode==PASTE){
-			System.out.println("Paste function not done yet");
+			updateReg.pasteSavedAtoms(updateReg.getViewer().getDisplayModelIndex());
 		}
-		else
-			System.out.println("Function copied");
+		else{
+			System.out.println("Invalid Operation");
+		}
+		//reset model atom seq number will give weird molecule.
+		/*for(Atom atm1 : updateReg.getModelList().get(currentModelIndex).getAtomHash().values()){
+			for(int i=0;i<atomNo.length;i++){
+				if(atm1.getAtomSeqNum()>atomNo[i])
+					atm1.setAtomSeqNum((atm1.getAtomSeqNum()-1));
+			}
+		}*/
 	}
 }
